Crate riddle_renderer_wgpu

Riddle simple sprite-based renderer built on wgpu.

The renderer can be used by itself as the only renderer in your application, or it can be constructed on top of existing WGPU devices to provide simple 2D rendering.

Riddle Example

The recommended way to use this crate is through the main riddle crate. Riddle exposes this crate through riddle::renderer. The following example lets this crate take full ownership over creating the WGPU device.

use riddle::{common::Color, platform::*, renderer::*, *};

fn main() -> Result<(), RiddleError> {
    let rdl =  RiddleLib::new()?;
    let window = WindowBuilder::new().build(rdl.context())?;
    let renderer = Renderer::new_from_window(&window)?;

    rdl.run(move |rdl| {
        match rdl.event() {
            Event::Platform(PlatformEvent::WindowClose(_)) => rdl.quit(),
            Event::ProcessFrame => {
                renderer.render(|render_ctx| {
                    render_ctx.clear(Color::RED)
                }).unwrap();
            },
            _ => ()
         }
    })
}

Direct Usage

To use this crate directly see WGPUDevice for how to use the renderer with custom WGPU devices.

Modules

vertex

Structs

Renderable

Renderer
A simple 2D sprite based renderer.

Sprite
A renderable region of a texture.

SpriteAtlasBuilder
Construct a set of Sprites from a set of riddle_image::Images which share a texture atlas.

SpriteFont
An efficient [riddle_font::ImgFont] renderer.

SpriteInitArgs

SpriteRenderArgs

SpriteRenderTarget
A target which can be both rendered to and referenced as a Sprite for rendering.

WGPUDeviceInfo

WindowWGPUDevice
A standard Renderer WGPU device used with riddle_platform_winit windows.

Enums

FilterMode

RendererError

WGPURendererError

Traits

CommonRenderer
The root object of a renderer implementation, associated with a single display.

CommonSprite
Sprites are conceptually both a reference to an image, and the sub region of the image which represents the logical sprite.

RenderContext
The context provided to render callbacks

WGPUDevice
A Renderer compatible WGPU device.

Type Definitions

DefaultRenderer