Struct riddle_renderer_wgpu::Sprite

pub struct Sprite<Device: WGPUDevice> { /* fields omitted */ }

A renderable region of a texture.

Multiple sprites can share a single texture. Sprites can either be built using Sprite, or SpriteAtlasBuilder.

Use SpriteRenderArgs for access to all supported paramters when rendering sprites, or use Sprite::render_at to specify only a location and use default arguments for everything else.

Sprites store a reference to the Renderer which built it, which will keep the renderer alive as long as the sprite is alive.

Example

let renderer = Renderer::new_from_window(&window)?;

// Load an image and create a sprite from it
let png_bytes = include_bytes!("../../example_assets/image.png");
let img = Image::load(&png_bytes[..], ImageFormat::Png)?;
let sprite = Sprite::new_from_image(&renderer, &img, &SpriteInitArgs::new())?;

// Render the sprite at the top left corner of the screen
let mut render_ctx = renderer.render(|render_ctx| {
    render_ctx.clear(Color::WHITE);
    sprite.render_at(render_ctx, vec2(0.0, 0.0))
})?;