Trait riddle_renderer_wgpu::RenderContext

pub trait RenderContext<R> where
    R: CommonRenderer, {
    pub fn set_transform(
        &mut self, 
        transform: ColumnMatrix4<f32>
    ) -> Result<(), RendererError>;

    pub fn clear(&mut self, color: Color<f32>) -> Result<(), RendererError>;

    pub fn draw(
        &mut self, 
        renderable: &Renderable<'_, R>
    ) -> Result<(), RendererError>;

    pub fn fill_rect(
        &mut self, 
        rect: &Rect<f32>, 
        color: Color<f32>
    ) -> Result<(), RendererError>;

    pub fn present(self) -> Result<(), RendererError>;
}

The context provided to render callbacks

Required methods

pub fn set_transform(
&mut self,
transform: ColumnMatrix4<f32>
) -> Result<(), RendererError>

Replace the current world transform.

pub fn clear(&mut self, color: Color<f32>) -> Result<(), RendererError>

Fill the target with a flat color.

pub fn draw(
&mut self,
renderable: &Renderable<'_, R>
) -> Result<(), RendererError>

Draw a Renderable to the target with the current world transform.

pub fn fill_rect(
&mut self,
rect: &Rect<f32>,
color: Color<f32>
) -> Result<(), RendererError>

Draw a solid rect with the given color.

pub fn present(self) -> Result<(), RendererError>

Consume the context and present any outstanding draw calls.

Loading content...