Module riddle_renderer_wgpu::wgpu_ext

Traits and structs required to use custom WGPU devices.

If using Renderer by itself through riddle there should be no need to use this module.

Structs

WGPUDeviceInfo

WGPURenderer
A simple 2D sprite based renderer.

WGPURendererHandle

WGPURendererWeak

WGPUSprite
A renderable region of a texture.

WGPUSpriteAtlasBuilder
Construct a set of WGPUSprites from a set of riddle_image::Images which share a texture atlas.

WGPUSpriteRenderTarget
A target which can be both rendered to and referenced as a Sprite for rendering.

WindowWGPUDevice
A standard Renderer WGPU device used with riddle_platform_winit windows.

Enums

FilterMode
Interpolation mode between texels when rendering

Traits

RenderContext
Types which accept render calls, tracks current world transform, and are consumed when the calls are presented.

WGPUDevice
A WGPURenderer compatible WGPU device.